+++ src/vmm/src/builder.rs
@@ -41,6 +41,8 @@ use crate::devices::virtio::vsock::{Vsock, VsockUnixBackend};
 use crate::gdb;
 use crate::initrd::{InitrdConfig, InitrdError};
 use crate::logger::debug;
+#[cfg(target_arch = "aarch64")]
+use crate::logger::warn;
 use crate::persist::{MicrovmState, MicrovmStateError};
 use crate::resources::VmResources;
 use crate::seccomp::BpfThreadMap;
@@ -195,6 +197,7 @@ pub fn build_microvm_for_boot(
         &vcpus_exit_evt,
         &vm,
         vm_resources.serial_out_path.as_ref(),
+        vm_resources.serial_rate_limiter(),
     )?;
 
     let vm = Arc::new(vm);
@@ -285,6 +288,7 @@ pub fn build_microvm_for_boot(
         event_manager,
         &mut boot_cmdline,
         vm_resources.serial_out_path.as_ref(),
+        vm_resources.serial_rate_limiter(),
     )?;
 
     device_manager.attach_vmgenid_device(&vm)?;
@@ -294,7 +298,7 @@ pub fn build_microvm_for_boot(
     if vcpus[0].kvm_vcpu.supports_pvtime() {
         setup_pvtime(&mut vm.resource_allocator(), &mut vcpus)?;
     } else {
-        log::warn!("Vcpus do not support pvtime, steal time will not be reported to guest");
+        warn!("Vcpus do not support pvtime, steal time will not be reported to guest");
     }
 
     configure_system_for_boot(