+++ src/vmm/src/device_manager/persist.rs
@@ -6,7 +6,6 @@
 use std::fmt::{self, Debug};
 use std::sync::{Arc, Mutex};
 
-use log::warn;
 use serde::{Deserialize, Serialize};
 
 use super::acpi::ACPIDeviceManager;
@@ -38,6 +37,7 @@ use crate::devices::virtio::vsock::persist::{
     VsockConstructorArgs, VsockState, VsockUdsConstructorArgs,
 };
 use crate::devices::virtio::vsock::{Vsock, VsockUnixBackend};
+use crate::logger::warn;
 use crate::mmds::data_store::MmdsVersion;
 use crate::resources::VmResources;
 use crate::snapshot::Persist;
@@ -373,6 +373,7 @@ impl<'a> Persist<'a> for MMIODeviceManager {
                         constructor_args.event_manager,
                         constructor_args.vm_resources.serial_out_path.as_ref(),
                         serial_state.as_ref(),
+                        constructor_args.vm_resources.serial_rate_limiter(),
                     )?;
 
                     dev_manager.register_mmio_serial(vm, serial, Some(state.device_info))?;