+++ src/vmm/src/signal_handler.rs
@@ -4,10 +4,9 @@
 use libc::{
     SIGBUS, SIGHUP, SIGILL, SIGPIPE, SIGSEGV, SIGSYS, SIGXCPU, SIGXFSZ, c_int, c_void, siginfo_t,
 };
-use log::error;
 
 use crate::FcExitCode;
-use crate::logger::{IncMetric, METRICS, StoreMetric};
+use crate::logger::{IncMetric, METRICS, StoreMetric, error_unrestricted};
 use crate::utils::signal::register_signal_handler;
 
 // The offset of `si_syscall` (offending syscall identifier) within the siginfo structure
@@ -23,7 +22,7 @@ const SYS_SECCOMP_CODE: i32 = 1;
 fn exit_with_code(exit_code: FcExitCode) {
     // Write the metrics before exiting.
     if let Err(err) = METRICS.write() {
-        error!("Failed to write metrics while stopping: {}", err);
+        error_unrestricted!("Failed to write metrics while stopping: {}", err);
     }
     // SAFETY: Safe because we're terminating the process anyway.
     unsafe { libc::_exit(exit_code as i32) };
@@ -43,9 +42,10 @@ macro_rules! generate_handler {
             }
             $signal_metric.store(1);
 
-            error!(
+            error_unrestricted!(
                 "Shutting down VM after intercepting signal {}, code {}.",
-                si_signo, si_code
+                si_signo,
+                si_code
             );
 
             $body(si_code, info);
@@ -67,7 +67,7 @@ fn log_sigsys_err(si_code: c_int, info: *mut siginfo_t) {
     // SAFETY: Other signals which might do async unsafe things incompatible with the rest of this
     // function are blocked due to the sa_mask used when registering the signal handler.
     let syscall = unsafe { *(info as *const i32).offset(SI_OFF_SYSCALL) };
-    error!(
+    error_unrestricted!(
         "Shutting down VM after intercepting a bad syscall ({}).",
         syscall
     );
@@ -141,13 +141,13 @@ extern "C" fn sigpipe_handler(num: c_int, info: *mut siginfo_t, _unused: *mut c_
     let si_code = unsafe { (*info).si_code };
 
     if num != si_signo || num != SIGPIPE {
-        error!("Received invalid signal {}, code {}.", si_signo, si_code);
+        error_unrestricted!("Received invalid signal {}, code {}.", si_signo, si_code);
         return;
     }
 
     METRICS.signals.sigpipe.inc();
 
-    error!("Received signal {}, code {}.", si_signo, si_code);
+    error_unrestricted!("Received signal {}, code {}.", si_signo, si_code);
 }
 
 /// Registers all the required signal handlers.