+++ tests/host_tools/fcmetrics.py
@@ -182,6 +182,7 @@ def validate_fc_metrics(metrics):
             "missed_metrics_count",
             "metrics_fails",
             "missed_log_count",
+            "rate_limited_log_count",
         ],
         "mmds": [
             "rx_accepted",
@@ -264,6 +265,7 @@ def validate_fc_metrics(metrics):
             "missed_write_count",
             "read_count",
             "write_count",
+            "rate_limiter_dropped_bytes",
         ],
         "signals": [
             "sigbus",